Fixed a 64-bit issue with the new NSDecimal rendering routines.

+++ examples/CPTestApp/Controller.h
@@ -18,7 +18,7 @@
 
 // Layer exploding for illustration
 -(IBAction)explodeLayers:(id)sender;
-+(void)recursivelySplitSublayersInZForLayer:(CALayer *)layer depthLevel:(unsigned int)depthLevel;
++(void)recursivelySplitSublayersInZForLayer:(CALayer *)layer depthLevel:(NSUInteger)depthLevel;
 -(IBAction)reassembleLayers:(id)sender;
 +(void)recursivelyAssembleSublayersInZForLayer:(CALayer *)layer;
